Fix USB errors due to unconfigured interrupt in endpoint
Although the subset of USB HID protocol that is actually used by the
bootloader uses only the default control endpoint (0), the HID spec
requires the device to have an interrupt in endpoint, and the host can
poll that endpoint even when the HID report descriptor does not actually
declare any input reports.  Polling an unconfigured endpoint causes USB
errors, which may prevent the bootloader from functioning properly.
Apparently this was happening in Mac OS, making the bootloader unusable
there (however, Windows and Linux did not expose the problem; on Linux
it was possible to provoke these errors by opening the `/dev/hidrawN`
device corresponding to the bootloader, but existing flashing tools do
not use that method to access the bootloader device).

Add the code to configure endpoint 1 as Interrupt IN, matching the USB
descriptors; this is enough to make the USB controller generate NAK
replies for that endpoint correctly, and the rest of bootloader code may
continue using just endpoint 0 as before.

The binary size increases by 12 bytes.

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>