Add bootloader overwrite protection
To protect the user from accidents (or even deliberate attempts to brick
the device), compare the specified page address with the bootloader
start address, and skip the flash write if the address could overlap
with the bootloader.

Handling the error case by jumping to `finish_hid_request` did not work
(it returned an error, but then the bootloader stopped responding to any
further USB requests); apparently it is important to consume the proper
number of bytes from the USB FIFO.  Because of that, the bad address
case is handled by running the same loop that is used for the normal
flash write case, but with all `spm` instructions disabled by setting a
flag bit.  Bit 7 of `reg_bRequest` is chosen for that role (that bit is
guaranteed to be 0 when starting to handle a normal `SET_HID_REPORT`
request without needing to add any instructions to clear it).

The `START_APPLICATION` command handling is also changed to use the same
code path to save space.

Because of additional optimization of the address comparison and the
`START_APPLICATION` handling code, the binary size is increased by just
2 bytes.

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>