Save 2 bytes in the DEVICE_TO_HOST parsing code
The DEVICE_TO_HOST handling code needs to handle only the GET_DESCRIPTOR
requests, which may come with two possible bmRequestType values:

  - 0x80 (IN direction, USB Standard Request, Recipient is the device) if
    a descriptor which applies to the device as a whole is requested (this
    code is used for the device and configuration descriptors);

  - 0x81 (IN direction, USB Standard Request, Recipient is the interface)
    if a descriptor specific to a particular interface is requested (this
    code is used for the HID class and HID report descriptors).

Because these codes are numerically sequential, and the direction bit has
already been tested (therefore it is known that bmRequestType >= 0x80), it
is enough to make a single comparison with 0x82 to detemine whether
bmRequestType has one of the above values.  Removing the bit masking
operation which is not needed after that change saves 2 bytes.