Save 6 bytes by refactoring UNHANDLED_SETUP_REQUEST usage, size 470 bytes
Remove the code which tried to check whether the SETUP packet has been
handled - now UNHANDLED_SETUP_REQUEST can be used only if the SETUP
packed needs to be acknowledged and replied with STALL, and in other
cases the code should just do `reti` directly.  In addition,
UNHANDLED_SETUP_REQUEST is placed in the middle of the code, so that it
would be reachable by conditional branches directly.

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>

This is based on
6bc21dd ("Save 6 bytes by refactoring UNHANDLED_SETUP_REQUEST usage", 2021-09-04)

Adjusted for thunk entry choice differences, both of them go to
UNHANDLED_SETUP_REQUEST:

 * UNHANDLED_SETUP_REQUEST_1
 * UNHANDLED_DEVICE_TO_HOST

Signed-off-by: Osamu Aoki <osamu@debian.org>